kernel-module-mcc: Only include mvf_sema4.h if building in kernel
When mcc_config.h is used to make application we shouldn't include
mvf_sema4.h as it is an internal symbol of kernel and shouldn't be
directly referenced by user space applications using the MCC
interface.

This allows for application developers to use the SDK without having
to copy the 'mvf_sema4.h' header byhand.

More detailed information at:

https://lists.yoctoproject.org/pipermail/meta-freescale/2015-January/012144.html

Change-Id: I39a519d10a15b951c573007d3cf689f32aac3f30
Reported-by: Petr Kubizňák <kubiznak.petr@elnico.cz>
Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>